+++ CHANGELOG.md
@@ -1,6 +1,45 @@
 CHANGELOG
 ==============
 
+2.1.1
+-----------------
+ * Code style
+ * Change src dir
+ * Added upload settings for site map
+
+2.1.0
+-----------------
+ * Using composer-config-plugin
+ 
+2.0.0.2
+-----------------
+  * Updated
+  
+2.0.0.1
+-----------------
+  * Disabled to backend
+  
+2.0.0
+-----------------
+  * Add registerLinkTags
+  * Add canonicalPageParams
+  
+2.0.0-beta
+-----------------
+  * Fixed errors
+  
+2.0.0-alpha
+-----------------
+  * SkeekS CMS v4
+  
+1.0.1
+-----------------
+  * Tree absolute url
+
+1.0.0
+-----------------
+  * Stable release
+ 
 1.0.0-alpha
 -----------------
   * Removed skeeks/cms/base/Module